Add movie controller
